+++ Code/Nielsen/Source/Events/NielsenReporter/AdEventReporter.swift
@@ -6,6 +6,7 @@
 //
 
 import THEOplayerSDK
+
 #if os(iOS)
 import NielsenAppApi
 #elseif os(tvOS)